+++ ConfigoSDK/ConfigoSDK/Classes/Controllers/location-controller/CFGLocationController.h
@@ -1,0 +1,15 @@
+//
+//  CFGLocationController.h
+//  ConfigoSDK
+//
+//  Created by Natan Abramov on 14/07/2016.
+//  Copyright © 2016 Configo. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface CFGLocationController : NSObject
+
+- (NSDictionary *)getLocationDictionary;
+
+@end