+++ ConfigoSDK/ConfigoSDKTests/OHHTTPStubsUtilities.m
@@ -47,7 +47,7 @@
 #pragma mark - Helpers
 
 + (void)setupStubLogs {
-    [OHHTTPStubs onStubActivation:^(NSURLRequest *request, id<OHHTTPStubsDescriptor> stub) {
+    [OHHTTPStubs onStubActivation:^(NSURLRequest *request, id<OHHTTPStubsDescriptor> stub, OHHTTPStubsResponse *response) {
         NSLog(@"%@ stubbed by %@.", request.URL, stub.name);
     }];
 }